requirements.txt
requests
miniwdl
six
